--- threadsafe_iter.py
@@ -1,16 +1,0 @@
-import threading
-#Credit: https://anandology.com/blog/using-iterators-and-generators/
-class threadsafe_iter:
-    """Takes an iterator/generator and makes it thread-safe by
-    serializing call to the `next` method of given iterator/generator.
-    """
-    def __init__(self, it):
-        self.it = it
-        self.lock = threading.Lock()
-
-    def __iter__(self):
-        return self
-
-    def next(self):
-        with self.lock:
-            return self.it.next()
